core/requirements.txt
Django==5.2.7
channels==4.3.1
channels-redis==4.3.0
Pillow==12.0.0
django-crispy-forms==2.4
crispy-tailwind==0.5.0
redis==7.0.0
asgiref==3.10.0
sqlparse==0.5.3
tzdata==2025.2
